requirements.txt
PyMuPDF
pdfplumber
pytesseract
Pillow
fastapi
uvicorn[standard]
python-multipart